+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 - PR #5673 Always build and test with per-thread default stream enabled in the GPU CI build
 - PR #5729 Create nvtext normalize_characters API from the subword_tokenize internal function
 - PR #5572 Add `cudf::encode` API.
+- PR #5568 Add support for `Series.keys()` and `DataFrame.keys()`
 
 ## Improvements